DATA_Sindex_2018_2023
Data set from the uBMS in BCN with the index of abundance calculated (based on the rbms package). Phenologies were set using uBMS raw counts and a subset of the CBMS raw counts with transects in the same bioclimatic region as the city + max altitude 600m (Criteria done by YM, DGC and CS).